apps/api/tests/test_health.py
import pytest

def test_health_check_basic():
    """
    Simple smoke test to ensure pytest is configured correctly.
    """
    assert 1 == 1

@pytest.mark.django_db
def test_database_access():
    """
    Verify database access is working (requires DB container up,
    but we mock or assume standard pytest-django setup).
    Since we don't have migrations run yet, we just check non-db logic or 
    basic model instantiation if possible. For now, just a pass.
    """
    pass
