apps/api/requirements.txt
django>=5.0
djangorestframework
psycopg2-binary
python-decouple
pytest-django
factory_boy
